added calloc strdup strjoin strncmp and notes and c hanged unsigned longs to size_t now that i included stdlib in the header

+++ libft/ft_strlcpy.c
@@ -6,13 +6,13 @@
 /*   By: echoukri <echoukri@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/12 18:32:20 by echoukri          #+#    #+#             */
-/*   Updated: 2022/10/14 12:12:21 by echoukri         ###   ########.fr       */
+/*   Updated: 2022/10/14 13:07:37 by echoukri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft.h"
 
-unsigned long ft_strlcpy(char * restrict dst, const char * restrict src, unsigned long dstsize)
+size_t ft_strlcpy(char * restrict dst, const char * restrict src, size_t dstsize)
 {
     unsigned int	index;
