+++ libft/libft.h
@@ -14,6 +14,7 @@
 #ifndef LIBFT_H
 #define LIBFT_H
 
+#include <stdlib.h>
 extern void	ft_bzero(void *s, size_t n);
 extern int	ft_toupper(int c);
 extern int	ft_tolower(int c);
@@ -26,8 +27,8 @@ extern int	ft_isdigit(int c);
 extern int	ft_isascii(int c);
 extern int	ft_isalpha(int c);
 extern int	ft_isalnum(int c);
-extern unsigned long ft_strlcpy(char * restrict dst, const char * restrict src, unsigned long dstsize);
+extern size_t ft_strlcpy(char * restrict dst, const char * restrict src, size_t dstsize);
 extern int	ft_atoi(char *str);
-extern unsigned long	ft_strlcat(char *dst, char *src, unsigned long size);
+extern size_t	ft_strlcat(char *dst, char *src, size_t size);
 
 #endif
